Fix shooting command
